+++ packages/openapi-generator/src/generator.ts
@@ -70,8 +70,11 @@ export async function generateWithParsedOptions(
   }
 
   if (options.clearOutputDir) {
+    // function rm was added in node version 14 and is the preferred method to use.
     const rm = promisesFs.rm || promisesFs.rmdir;
-    await rm(options.outputDir, { recursive: true });
+    const forceOption =
+      typeof promisesFs.rm === 'undefined' ? {} : { force: true };
+    await rm(options.outputDir, { recursive: true, ...forceOption });
   }
   const inputFilePaths = await getInputFilePaths(options.input);