+++ test-packages/self-tests/dtslint-test/error-test.spec.ts
@@ -1,4 +1,0 @@
-// This test is meant to fail in order to check if type tests are working.
-
-// $ExpectType number
-'0123'.charAt(0);