+++ packages/mail-client/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @sap-cloud-sdk/mail-client
 
+## 2.9.0
+
+### Minor Changes
+
+- d1bf2dee: [New Functionality] Expose SMTP transport options of `nodemailer`.
+
+### Patch Changes
+
+- Updated dependencies [4c51d3dc]
+- Updated dependencies [24029503]
+  - @sap-cloud-sdk/util@2.9.0
+  - @sap-cloud-sdk/connectivity@2.9.0
+
 ## 2.8.0
 
 ### Minor Changes